fix intermediate api use
